Implementato servizio che aggiorna lo stato Per ora aggiorna solamente Home
