+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.unison.appartment">
+
     <!-- Permesso per poter usare il registratore -->
     <uses-permission android:name="android.permission.RECORD_AUDIO" />
 
@@ -12,7 +13,12 @@
         android:roundIcon="@mipmap/ic_launcher_appartment_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".activities.UserProfileActivity"></activity>
+        <service
+            android:name=".AppartmentService"
+            android:enabled="true"
+            android:exported="true"></service>
+
+        <activity android:name=".activities.UserProfileActivity" />
         <activity android:name=".activities.JoinHomeActivity" />
         <activity android:name=".activities.ImageDetailActivity" />
         <activity android:name=".activities.RewardDetailActivity" />
